requirements.txt
streamlit==1.51.0
requests==2.32.5
pandas==2.3.3
plotly==6.4.0
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0